# Use octree traversal to implement star browser functions