- Store node depth instead of size in static octree
- Implement "breadth-first" octree traversal with iterative depth-first processing
- Rewrite the star browser comparison functions to use breadth-first traversal