@375gnu 2024-10-09
Could you briefly describe advantages of the new implementation (preferably in the commit message)?