@ajtribick 2024-10-09
Performance measurement from near to Sol in the Qt star browser (average of 5 refreshes):

**Closest stars**

Old: 22193040 ns
New: 2503289 ns (8.9× faster)

**Closest stars with planets**

Old: 135039170 ns
New: 55153973 ns (2.4× faster)

**Brightest stars**

Old: 65632337 ns
New: 3158870 ns (20.8× faster)

**Brightest stars with planets**

Old: 134337381 ns
New: 19497552 ns (6.9× faster)